0.2.3
    - Cebullus located and fixed item modification error with weights
    - Added new Learning limit multipliers added
Added new strings
    <string id="KTMCM_XPMLL" text="Learning Limit Enable" />
    <string id="KTMCM_XPMLLH" text="Enable Learning Limit modifiers." />
    <string id="KTMCM_XPMLLM" text="Learning Limit Multiplier" />
    <string id="KTMCM_XPMLLMH" text="Multiply Learning limit by the multiplier [Native : 0.0[0%]]." />
    <string id="KTMCM_CLearningLimit" text="Learning Limit" />

+++ KaosesTweaks/Objects/Items/Armor.cs
@@ -31,7 +31,7 @@ namespace KaosesTweaks.Objects
             }
             if (_settings.ItemArmorWeightModifiers && _settings.MCMArmorModifiers)
             {
-                SetItemsWeight((int)(_item.Weight * multiplerWeight), multiplerWeight);
+                SetItemsWeight(_item.Weight * multiplerWeight, multiplerWeight);
             }
         }
